Update PeriodoRepository.java
cambio a long

+++ src/main/java/com/example/backend/repositories/PeriodoRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repositories;
 import com.example.backend.entities.periodo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeriodoRepository extends JpaRepository<periodo, Integer> {
+public interface PeriodoRepository extends JpaRepository<periodo, Long> {
 }
